+++ packages/indid-core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @indid/indid-core-sdk
 
+## 0.0.3
+
+### Patch Changes
+
+- 8e7088f: fix: websocket url
+
+## 0.0.2
+
+### Patch Changes
+
+- e0d6c52: non sequential nonce
+- Updated dependencies [e0d6c52]
+  - @indid/indid-typechains@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes